requirements.txt
Flask-RESTful
Flask-SQLAlchemy
PyJWT
alembic
Flask-Migrate
sqlalchemy-utils
gunicorn
psycopg2
googlemaps
passlib
sendgrid
requests